% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchLengthMetrics.R
\name{branchLengthMetrics}
\alias{branchLengthMetrics}
\title{Produces a dataframe of tree branch length statistics}
\usage{
branchLengthMetrics(tree)
}
\arguments{
\item{tree}{a phylogenetic tree (as a \code{phylo} object)}
}
\value{
TODO
}
\description{
\code{branchLengthMetrics} returns a list of imabalance statistics
}
\author{
John Lees (\email{lees.john6@gmail.com})
}

