% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{compute.distance.matrix}
\alias{compute.distance.matrix}
\title{Computes a pairwise distance matrix}
\usage{
compute.distance.matrix(aln, model = "TN93", gamma = FALSE)
}
\arguments{
\item{aln}{a multiple sequence alignment}

\item{model}{a model of nucleotide substitution (default=TN93)}

\item{gamma}{the gamma parameter of rate variation}
}
\value{
A distance matrix (as a standard R matrix)
}
\description{
\code{compute.distance.matrix} calculates a pairwise distance matrix based on a multiple sequence
alignment. It is basically a thin wrapper to \code{dist.dna} in \code{ape}, with the default model
being TN93 rather than K80 and pairwise deletions considered, which returns a matrix
(rather than either a \code{dist} object or a matrix).
}
\author{
Simon Frost (\email{sdwfrost@gmail.com})
}

