% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{expand.distance.matrix}
\alias{expand.distance.matrix}
\title{Updates a pre-existing distance matrix}
\usage{
expand.distance.matrix(aln, s, distmat = NULL, model = "TN93",
  gamma = FALSE)
}
\arguments{
\item{aln}{a multiple sequence alignment}

\item{s}{a nucleotide sequence}

\item{distmat}{a distance matrix (as a standard matrix). If \code{NULL}, then a distance matrix is computed.}

\item{model}{a model of nucleotide substitution (default=TN93)}

\item{gamma}{the gamma parameter of rate variation}
}
\value{
A list with two elements: distmat, a distance matrix (as a standard R matrix), and aln, an updated
alignment.
}
\description{
\code{expand.distance.matrix} updates a pairwise distance matrix with distances calculated between
the existing alignment and an additional sequence (assumed to be aligned)
}
\author{
Simon Frost (\email{sdwfrost@gmail.com})
}

