% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalanceMetrics.R
\name{imbalanceMetrics}
\alias{imbalanceMetrics}
\title{Produces a dataframe of tree imbalance statistics}
\usage{
imbalanceMetrics(tree)
}
\arguments{
\item{tree}{a phylogenetic tree (as a \code{phylo} object)}
}
\value{
A list of 14 imbalance metrics. TO DO. Improve local branching index utility
}
\description{
\code{imbalanceMetrics} returns a list of imabalance statistics
}
\author{
OJ Watson (\email{o.watson15@imperial.ac.uk})
}

