% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalanceThreshold.R
\name{imbalanceThreshold}
\alias{imbalanceThreshold}
\title{Produces a list of a confidence intervals given a tree through tip node topology permutation}
\usage{
imbalanceThreshold(tree, reps = 5, ci = 0.025)
}
\arguments{
\item{tree}{a phylogenetic tree (as a \code{phylo} object)}

\item{reps}{the number of permutations to execute for the tree}

\item{ci}{confidence interval. Default = 0.25, i.e. 95 percent}
}
\value{
A list of length 12 for th first 12 imbalance metrics
}
\description{
\code{imbalanceThreshold} returns a list of thresholds
}
\author{
OJ Watson (\email{o.watson15@imperial.ac.uk}) Yu Luo (\email{yu.luo3@mail.mcgill.ca})
}

