% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{pairwise.distance}
\alias{pairwise.distance}
\title{Calculate pairwise distance between two sequences}
\usage{
pairwise.distance(s1, s2, model = "TN93", gamma = FALSE)
}
\arguments{
\item{s1}{a nucleotide sequence}

\item{s2}{a second nucleotide sequence. The sequences should be of equal length}

\item{model}{model of nucleotide substitution (default=TN93)}

\item{gamma}{The gamma parameter of rate variation}
}
\value{
The distance (as a double).
}
\description{
\code{pairwise.distance} calculates the pairwise distance between two sequences.
}
\author{
Simon Frost (\email{sdwfrost@gmail.com})
}

