% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteTrees.R
\name{permuteTrees}
\alias{permuteTrees}
\title{Produces a list of permuted trees with the same tip and internal node times for bootstrapping purposes}
\usage{
permuteTrees(tree, reps = 100)
}
\arguments{
\item{tree}{a phylogenetic tree (as a \code{phylo} object)}

\item{reps}{the number of permuted trees}
}
\value{
A list of permuted trees
}
\description{
\code{permuteTrees} returns the length of a tip of a tree, given either (a) the tip name or
(b) the index of the tip (in the order of the tip labels in the tree).
}
\author{
OJ Watson (\email{o.watson15@imperial.ac.uk})
}

