% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBranchLengths.R
\name{plot.branch.lengths}
\alias{plot.branch.lengths}
\title{Plots the average branch lengths over time}
\usage{
plot.branch.lengths(tree, permutes = 10)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{permutes}{number of permutation (passed to prunePermute)}
}
\description{
\code{plot.branch.lengths} returns a list of imabalance statistics
}
\author{
John Lees (\email{lees.john6@gmail.com})
}

