% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prunePermute.R
\name{prunePermute}
\alias{prunePermute}
\title{Produces a list of a list of sequentially growing trees given a simulated tree.}
\usage{
prunePermute(tree, permutes = 10)
}
\arguments{
\item{tree}{a phylogenetic tree (as a \code{phylo} object)}

\item{permutes}{the number of permutations to execute for the tree}
}
\value{
A list of length equal to the number of nodes in the given tree
}
\description{
\code{prunePermute} returns a list of length equal to the number of nodes in the given tree
}
\author{
OJ Watson (\email{o.watson15@imperial.ac.uk})
}

