% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatetree.R
\name{simulatetree}
\alias{simulatetree}
\title{Simulate tree and sequences in two-population SIR-model}
\usage{
simulatetree(N = c(100, 10000), params = list(beta = matrix(c(2, 0, 0, 2),
  nrow = 2), gamma = 1, mu = 0.05), I0 = c(1, 1), t1 = 100, res = 1000,
  sampletimes = c("incidence", "regular"), samplesize = 200,
  sampleperiod = c(0, t1), sampledist = c("incidence", "popsize", "fixed"),
  samplebias = 1, ...)
}
\arguments{
\item{N}{the population sizes of the two populations}

\item{params}{the parameters for the SIR-model: a named list with transmission rate matrix \code{beta},
recovery rate \code{gamma} and death rate \code{mu}. Births balance deaths. Transmission is modelled with frequency-dependent
transmission terms (division by N).}

\item{I0}{initial numbers of infecteds. The rest is assumed susceptible.}

\item{t1}{duration of the simulation}

\item{res}{resolution of the epidemic simulation (ODE solver)}

\item{sampletimes}{indicates if sampling occurs at random times proportional to incidence (default \code{"incidence"}) or
at random uniform times (\code{"regular"})}

\item{samplesize}{the number of samples to be taken}

\item{sampleperiod}{the time interval from which the sampling times should be taken}

\item{sampledist}{indicates how samples should be divided over the two populations, apart from bias. The default \code{"incidence"} indicates
proportionality to incidence; \code{"popsize"} indicates proportionality to population size; \code{"fixed"} indicates
no proportionality, i.e. fifty-fifty}

\item{samplebias}{indicates the extra weight (multiplication factor) given for sampling the first population
relative to the second. If equal to 1, the weight is completely determined by \code{sampledist}.}

\item{...}{arguments to be passed to simSeq (\pkg{phangorn})}
}
\value{
A list with a tree (classes \code{phylo} and \code{DatedTree}), and a set of sequences (class \code{phyDat}).
}
\description{
\code{simulatetree} does a deterministic simulation of a two-population SIR-model,
takes samples and simulates a phylogenetic tree. For all this, it makes use of phydynR
(https://github.com/emvolz-phylodynamics/phydynR). Then it simulates DNA-sequences for these samples, with simSeq (\pkg{phangorn}).
}
\author{
Don Klinkenberg (\email{don@xs4all.nl})
}

