% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeprune.R
\name{timeprune}
\alias{timeprune}
\title{Prune tree in time order}
\usage{
timeprune(tr, tipdates = NULL, reverse = FALSE)
}
\arguments{
\item{tr}{a tree, as a \code{phylo} object}

\item{tipdates}{a vector of sampling times or dates}

\item{reverse}{a boolean (default: \code{FALSE}) determining whether the list of trees should be returned
in reverse time}
}
\value{
A list with tipdates and trees.
}
\description{
\code{timeprune} prunes a tree by removing tips one at a time, in reverse time order.
If the time is omitted and the tree is rooted, then the root-to-tip distance is used as
a proxy of sampling time.
}
\author{
Simon Frost (\email{sdwfrost@gmail.com})
}

